requirements.txt
luma.oled>=3.6.0
numpy
picamera[array]
opencv-contrib-python==4.1.0.25
